Update Intrument Importer sevice and rake task

+++ db/migrate/20241225101550_create_mis_details.rb
@@ -3,9 +3,9 @@ class CreateMisDetails < ActiveRecord::Migration[8.0]
     create_table :mis_details do |t|
       t.references :instrument, null: false, foreign_key: true
       t.string :isin
-      t.decimal :mis_leverage
-      t.decimal :bo_leverage
-      t.decimal :co_leverage
+      t.integer :mis_leverage
+      t.integer :bo_leverage
+      t.integer :co_leverage
 
       t.timestamps
     end
